requirements.txt
# Core dependencies
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.28.1
tqdm>=4.64.1

# Document processing
chardet>=5.0.0
PyPDF2>=3.0.0
pdfplumber>=0.7.6
PyMuPDF>=1.21.1
python-docx>=0.8.11 